mingw32 build: fix various warnings

+++ client/windows/event_sources_p.cpp
@@ -88,7 +88,7 @@ bool EventSources::wait_events(int timeout_ms)
         THROW("wait failed %d", GetLastError());
     }
 
-    int event_index = wait_res - WAIT_OBJECT_0;
+    size_t event_index = wait_res - WAIT_OBJECT_0;
     if (event_index == _handles.size()) {
         return process_system_events();
     } else if ((event_index >= 0) && (event_index < (int)_handles.size())) {
